CRUD Product Client, ErrorHandler Client

+++ src/components/ProductList.vue
@@ -43,6 +43,8 @@
 </template>
 
 <script>
+import Swal from 'sweetalert2'
+
 export default {
   name: 'ProductList',
   methods: {
@@ -50,7 +52,24 @@ export default {
       this.$store.dispatch('fetchProducts')
     },
     deleteProduct (id) {
-      this.$store.dispatch('deleteProduct', id)
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+          )
+          this.$store.dispatch('deleteProduct', id)
+        }
+      })
     },
     editProduct (id) {
       this.$router.push({ path: `products/edit/${id}` })
